Improve ProjectionRowEntry and Header inheritance from same style
